+++ src/modules/coordToolkit/tests/unit/store/gettersCoordToolkit.spec.js
@@ -3,10 +3,6 @@ import sinon from "sinon";
 import getters from "@modules/coordToolkit/store/gettersCoordToolkit.js";
 import toolKitState from "@modules/coordToolkit/store/stateCoordToolkit.js";
 
-afterEach(() => {
-    sinon.restore();
-});
-
 describe("src/modules/coordToolkit/store/gettersCoordToolkit.js", () => {
     let stateCoord;